split state/mod.rs into several files
